Update usb_busses state after supposed device disconnecting
The libusb-compat hasn't monitoring for disconnect, so usb_busses
keeps the state of connected devices unless next call to
usb_find_devices. If the device was disconnected without call to
usb_find_devices, after reconnecting the usb_find_devices will find the
device with the same address that it already has cached and it will not
update its handle. So it will lead to work with irrelevant device and
we will not be able to work with it.

Resolve: NANO-1372